requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
PyYAML>=6.0
yamllint>=1.29.0